requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyotp>=2.9.0
MetaTrader5>=5.0.45
websocket-client>=1.6.0
